+++ client/src/test/java/com/influxdb/client/ITAuthorizationsApi.java
@@ -38,7 +38,6 @@ import com.influxdb.exceptions.NotFoundException;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.platform.runner.JUnitPlatform;
 import org.junit.runner.RunWith;
@@ -309,8 +308,6 @@ class ITAuthorizationsApi extends AbstractITClientTest {
     }
 
     @Test
-    //TODO https://github.com/influxdata/influxdb/issues/16833
-    @Disabled
     void deleteAuthorization() {
 
         Authorization createdAuthorization = authorizationsApi.createAuthorization(organization, newPermissions());