+++ client/src/test/java/com/influxdb/client/ITNotificationEndpointsApi.java
@@ -42,6 +42,7 @@ import com.influxdb.exceptions.NotFoundException;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.platform.runner.JUnitPlatform;
 import org.junit.runner.RunWith;
@@ -584,6 +585,8 @@ class ITNotificationEndpointsApi extends AbstractITClientTest {
                 .isInstanceOf(NotFoundException.class);
     }
 
+    //TODO wait to v2.0.0-beta.7
+    @Disabled
     @Test
     void labelDeleteNotExists() {
 
@@ -591,7 +594,9 @@ class ITNotificationEndpointsApi extends AbstractITClientTest {
                 .createHTTPEndpointBearer(generateName("http"), "http://localhost:1234/mock",
                         HTTPNotificationEndpoint.MethodEnum.GET, "my-token", orgID);
 
-        notificationEndpointsApi.deleteLabel("020f755c3c082000", endpoint.getId());
+        Assertions.assertThatThrownBy(() ->notificationEndpointsApi.deleteLabel("020f755c3c082000", endpoint.getId()))
+                .isInstanceOf(NotFoundException.class)
+                .hasMessage("label not found");
     }
 
     @Test